Move log stuff to top level
